refactor(#5071): isolate sink terminal handoff
Move the typed sink ACK fold and RelaySink delivery orchestration behind a focused module boundary without changing delivery behavior.
